Fixed version of libc; added more constant mappings

+++ src/bindgen/constants/ETHER.rs
@@ -2,14 +2,14 @@
 // Copyright © 2016 The developers of dpdk-sys. See the COPYRIGHT file in the top-level directory of this distribution and at https://raw.githubusercontent.com/lemonrock/dpdk-sys/master/COPYRIGHT.
 
 
-pub const ETHER_CRC_LEN: c_int = 4;
+pub const ETHER_CRC_LEN: uint16_t = 4;
 pub const ETHER_GROUP_ADDR: c_int = 1;
 pub const ETHER_HDR_LEN: c_int = 14;
 pub const ETHER_LOCAL_ADMIN_ADDR: c_int = 2;
-pub const ETHER_MAX_JUMBO_FRAME_LEN: size_t = 16128;
+pub const ETHER_MAX_JUMBO_FRAME_LEN: uint16_t = 16128;
 pub const ETHER_MAX_LEN: size_t = 1518;
 pub const ETHER_MAX_VLAN_FRAME_LEN: c_int = 1522;
 pub const ETHER_MAX_VLAN_ID: c_int = 4095;
 pub const ETHER_MIN_LEN: size_t = 64;
-pub const ETHER_MIN_MTU: c_int = 68;
-pub const ETHER_MTU: c_int = 1500;
+pub const ETHER_MIN_MTU: uint16_t = 68;
+pub const ETHER_MTU: uint16_t = 1500;
